+++ src/clicknick/__init__.py
@@ -1,3 +1,3 @@
-from .app import main, main_dev
+from .app import main
 
-__all__ = ["main", "main_dev"]
+__all__ = ["main"]